add new input format (png images directory with device timestamps)

+++ M2P_test_analysis/M2P_test_analysis.h
@@ -12,7 +12,7 @@ enum FRAME_CONTROL {
 };
 
 typedef struct {
-	int frameID;
+  double frameID;
 	cv::Point2f uv;
 	cv::Point2f realPositionInPixel;
 	cv::Point2f expectedPositionInPixel;
